+++ lib/new_relic/agent/configuration/default_source.rb
@@ -916,6 +916,14 @@ If `true`, disables agent middleware for Sinatra. This middleware is responsible
           :allowed_from_server => false,
           :description => 'Controls auto-instrumentation of bunny at start up.  May be one of [auto|prepend|chain|disabled].'
         },
+        :'instrumentation.elasticsearch' => {
+          :default => 'auto',
+          :public => true,
+          :type => String,
+          :dynamic_name => true,
+          :allowed_from_server => false,
+          :description => 'Controls auto-instrumentation of the elasticsearch library at start up. May be one of [auto|prepend|chain|disabled].'
+        },
         :'instrumentation.httprb' => {
           :default => instrumentation_value_of(:disable_httprb),
           :documentation_default => 'auto',
@@ -1406,6 +1414,20 @@ If `true`, disables agent middleware for Sinatra. This middleware is responsible
           :type => Boolean,
           :allowed_from_server => true,
           :description => 'If `true`, the agent obfuscates Mongo queries in transaction traces.'
+        },
+        :'elasticsearch.capture_queries' => {
+          :default => true,
+          :public => true,
+          :type => Boolean,
+          :allowed_from_server => true,
+          :description => 'If `true`, the agent captures Elasticsearch queries in transaction traces.'
+        },
+        :'elasticsearch.obfuscate_queries' => {
+          :default => true,
+          :public => true,
+          :type => Boolean,
+          :allowed_from_server => true,
+          :description => 'If `true`, the agent obfuscates Elasticsearch queries in transaction traces.'
         },
         :'error_collector.enabled' => {
           :default => true,